Location Refactor prevent multiple presses when "loading"
