+++ src/config.jsx
@@ -1,3 +1,5 @@
 const apiUrl = 'https://mobility-mate-production.up.railway.app';
+// const apiUrl = 'http://127.0.0.1:5000';
+
 // console.log(process.env.REACT_APP_FLASK_API_URL);
 export default apiUrl;
